AaltoFloatsEdge/modules/logmodule/requirements.txt
azure-iot-device~=2.7.0
flatten-json~=0.1.13
 azure-storage-blob~=12.12.0